+++ model/action.gen.go
@@ -3,9 +3,9 @@ package model
 const TabelAction = "action"
 
 type Action struct {
-	ID        int32  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
+	ID        int64  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
 	Type      string `gorm:"column:type;not null" json:"type"`
-	TriggerID int32  `gorm:"column:triggerId;not null" json:"triggerId"`
+	TriggerID int64  `gorm:"column:triggerId;not null" json:"triggerId"`
 	WaitTime  string `gorm:"column:wait_time;not null" json:"waitTime"`
 }
