+++ model/customersTocustomer_groups.go
@@ -9,10 +9,10 @@ const TableNameItemsToitemCategory = "_CustomerTocustomer_groups"
 // ItemsToitemCategory mapped from table <_ItemsToitem_category>
 type CustomersToCustomerGroups struct {
 	// A is a foreign key to Customer table that has primary key ID
-	A int32 `gorm:"column:A" json:"A"`
+	A int64 `gorm:"column:A" json:"A"`
 
 	// B is a foreign key to CustomerGroup table that has primary key ID
-	B int32 `gorm:"column:B" json:"B"`
+	B int64 `gorm:"column:B" json:"B"`
 
 	// many to one relationship of Customer to CustomerGroup
 	Customer Customer `gorm:"foreignKey:A;references:ID"`
