+++ model/sms_action.gen.go
@@ -3,9 +3,9 @@ package model
 const TabelSms_action = "sms_action"
 
 type SmsAction struct {
-	ID       int32  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
+	ID       int64  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
 	Body     string `gorm:"column:body;not null" json:"body"`
-	ActionID int32  `gorm:"column:actionId;not null" json:"actionId"`
+	ActionID int64  `gorm:"column:actionId;not null" json:"actionId"`
 }
 
 // TableName FbTarget's table name
